+++ src/modules/duels/components/css/board.css
@@ -31,7 +31,6 @@ img {
   cursor: pointer;
   border: 2px solid black;
   overflow: hidden;
-  margin-top: 5px;
 }
 
 .mini-carta {
