+++ src/modules/duels/components/selectionMode.jsx
