+++ src/modules/duels/components/v1components/revealCard.jsx
