+++ src/modules/duels/components/v2Components/basicCard.jsx
