+++ src/modules/duels/components/v2Components/board.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import "../components/css/board.css";
+import "../css/board.css";
 import LifeContainer from "./lifeArea";
 import CharacterArea from "./characterArea";
 import LeaderArea from "./leaderArea";
