+++ src/modules/duels/components/v2Components/duelCard.jsx
