+++ src/modules/duels/components/v2Components/lifeArea.jsx
