+++ src/modules/duels/components/v2Components/rivalBoard.jsx
