+++ src/modules/duels/components/vsPlayer.jsx
@@ -1,14 +1,13 @@
 import React from "react";
 import PreviewAndPhaseZone from "../components/previewAndPhasesArea/previewAndPhaseZone";
-import Board from "./board";
-import RivalBoard from "./rivalBoard";
+import Board from "../components/v1components/board";
+import BoardV2 from "../components/v2Components/board";
 
 function VsPlayer() {
   return (
     <div className="full-board bg-dark">
       <PreviewAndPhaseZone />
       <div>
-        <RivalBoard/>
         <Board />
       </div>
     </div>
