+++ src/modules/duels/views/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Store from "../provider/duelProvider";
-import Mode from "./selectionMode";
+import Mode from "../components/selectionMode";
 import Vs from "./vs";
 import { useLocation } from "react-router-dom";
 import "../components/css/duel.css";
